Finished startup and added feedback form fragment
