+++ app/src/main/res/layout/fragment_new_event_details.xml
@@ -3,15 +3,14 @@
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 xmlns:app="http://schemas.android.com/apk/res-auto"
-xmlns:tools="http://schemas.android.com/tools"
-android:padding="16dp">
+xmlns:tools="http://schemas.android.com/tools">
 
 <TextView
     android:id="@+id/edit_event_header"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:padding="16dp"
-    android:background="@color/tea_green"
+    android:background="@color/periwinkle"
     android:textSize="25sp"
     android:textAlignment="center"
     android:text="Event Details"
@@ -26,7 +25,7 @@ android:padding="16dp">
     android:layout_width="wrap_content"
     android:layout_height="200dp"
     tools:srcCompat="@tools:sample/backgrounds/scenic"
-
+    android:padding="16dp"
     app:layout_constraintTop_toBottomOf="@id/edit_event_header"
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintEnd_toEndOf="parent" />
@@ -64,7 +63,8 @@ android:padding="16dp">
     android:layout_height="0dp"
     android:orientation="vertical"
     app:layout_constraintTop_toBottomOf="@id/edit_event_poster_view"
-    app:layout_constraintBottom_toBottomOf="parent">
+    app:layout_constraintBottom_toBottomOf="parent"
+    >
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -111,7 +111,8 @@ android:padding="16dp">
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textSize="20dp"
-            android:text="Require Geolocation"/>
+            android:text="Require Geolocation"
+            android:padding="16dp"/>
         <CheckBox
             android:id="@+id/view_checkbox_geolocation_requirement"
             android:enabled="false"
@@ -124,19 +125,21 @@ android:padding="16dp">
         android:layout_height="wrap_content"
         android:paddingTop="10dp"
         android:textSize="20dp"
-        android:text="Event Description:"/>
+        android:text="Event Description:"
+        android:padding="16dp"/>
 
     <TextView
         android:id="@+id/view_event_description"
         android:textSize="20dp"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content"
+        android:padding="16dp"/>
 
 </LinearLayout>
 
 <Button
     android:id="@+id/join_waitlist_button"
-    style="@style/ButtonStyleB"
+    style="@style/ButtonStyleA"
     app:layout_constraintTop_toTopOf="parent"
     app:layout_constraintBottom_toBottomOf="parent"
     app:layout_constraintStart_toStartOf="parent"
