+++ app/src/main/res/layout/fragment_waitlisted_event_details.xml
@@ -3,15 +3,14 @@
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 xmlns:app="http://schemas.android.com/apk/res-auto"
-xmlns:tools="http://schemas.android.com/tools"
-android:padding="16dp">
+xmlns:tools="http://schemas.android.com/tools">
 
 <TextView
     android:id="@+id/edit_event_header"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:padding="16dp"
-    android:background="@color/tea_green"
+    android:background="@color/periwinkle"
     android:textSize="25sp"
     android:textAlignment="center"
     android:text="Edit Event Details"
@@ -29,12 +28,12 @@ android:padding="16dp">
 
     app:layout_constraintTop_toBottomOf="@id/edit_event_header"
     app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintEnd_toEndOf="parent" />
+    app:layout_constraintEnd_toEndOf="parent"
+    android:padding="16dp"/>
 
 <ImageButton
     android:id="@+id/cancel_button"
     style="@style/XButtonStyle"
-
     app:layout_constraintTop_toTopOf="parent"
     app:layout_constraintBottom_toBottomOf="parent"
     app:layout_constraintStart_toStartOf="parent"
@@ -91,14 +90,16 @@ android:padding="16dp">
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textSize="20dp"
-            android:text="waitlist capacity: "/>
+            android:text="waitlist capacity: "
+            android:padding="16dp"/>
 
         <TextView
             android:id="@+id/view_waitlist_capacity"
             android:layout_width="100dp"
             android:textAlignment="center"
             android:textSize="20dp"
-            android:layout_height="wrap_content"/>
+            android:layout_height="wrap_content"
+            android:padding="16dp"/>
 
     </LinearLayout>
     <!--Require Geolocation Option-->
@@ -111,12 +112,14 @@ android:padding="16dp">
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textSize="20dp"
-            android:text="Require Entrant Geolocation"/>
+            android:text="Require Entrant Geolocation"
+            android:padding="16dp"/>
         <CheckBox
             android:id="@+id/view_checkbox_geolocation_requirement"
             android:enabled="false"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content"/>
+            android:layout_height="wrap_content"
+            android:padding="16dp"/>
     </LinearLayout>
 
     <TextView
@@ -124,13 +127,15 @@ android:padding="16dp">
         android:layout_height="wrap_content"
         android:paddingTop="10dp"
         android:textSize="20dp"
-        android:text="Event Description:"/>
+        android:text="Event Description:"
+        android:padding="16dp"/>
 
     <TextView
         android:id="@+id/view_event_description"
         android:textSize="20dp"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content"
+        android:padding="16dp"/>
 
 </LinearLayout>
