+++ app/src/main/res/layout/startup.xml
@@ -31,7 +31,7 @@
         android:id="@+id/linearLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="16dp"
         android:orientation="horizontal"
         android:padding="8dp"
         android:paddingBottom="8dp"
@@ -58,7 +58,7 @@
         android:id="@+id/linearLayout3"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="16dp"
         android:orientation="horizontal"
         android:padding="8dp"
         android:paddingBottom="8dp"
@@ -85,7 +85,7 @@
         android:id="@+id/linearLayout4"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="16dp"
         android:orientation="horizontal"
         android:padding="8dp"
         android:paddingBottom="8dp"
@@ -107,20 +107,92 @@
             android:layout_height="wrap_content"
             android:layout_weight="1"
             android:hint="Enter Phone (optional)" />
+
     </LinearLayout>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/organizer_startup_fields"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="24dp"
+        android:visibility="gone"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/linearLayout4">
+
+        <LinearLayout
+            android:id="@+id/linearLayout5"
+            android:layout_width="408dp"
+            android:layout_height="65dp"
+            android:orientation="horizontal"
+            android:padding="8dp"
+            android:paddingBottom="8dp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent">
+
+            <TextView
+                style="@style/FragmentConstants"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="User Facility:" />
+
+            <EditText
+                android:id="@+id/edit_user_facility"
+                style="@style/FragmentConstants"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:hint="Enter Facility (optional)" />
+
+        </LinearLayout>
+
+        <LinearLayout
+            android:id="@+id/linearLayout6"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="16dp"
+            android:orientation="horizontal"
+            android:padding="8dp"
+            android:paddingBottom="8dp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.0"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/linearLayout5">
+
+            <TextView
+                style="@style/FragmentConstants"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Facility Address:" />
+
+            <EditText
+                android:id="@+id/edit_user_facility_address"
+                style="@style/FragmentConstants"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:hint="Enter Facility Address (optional)" />
+
+        </LinearLayout>
+
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
     <Button
         android:id="@+id/startup_button"
         android:layout_width="162dp"
         android:layout_height="68dp"
-        android:layout_marginTop="32dp"
+        android:layout_marginBottom="24dp"
         android:text="Finish"
-        android:theme="@style/ButtonStyleA"
         android:textColor="@color/marian_blue"
+        android:theme="@style/ButtonStyleA"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/linearLayout4" />
+        app:layout_constraintVertical_bias="0.848" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
